+++ src/config.h
@@ -4,7 +4,7 @@
 extern double LONGITUDINAL_SMOOTHNESS;
 extern int SHAPE_CURVE_SAMPLES;
 extern double STRUCTURAL_MARGIN;
-extern double MESH_ALPHA;
+extern float MESH_ALPHA;
 extern float ONE_SIDE_MERGE_DELAY;
 extern float TWO_SIDE_MERGE_DELAY;
 extern float ONE_MINUS_ONE_SIDE_MERGE_DELAY;
